We have to change the stop handler tag to ensure that patient name id and birth date tags are available when we are checking for anonymisation. Added a null pointer check for closing the series zip file. Update to the project file

+++ src/main/java/uk/ac/ucl/cs/cmic/giftcloud/restserver/ZipSeriesRequestFixedSize.java
@@ -81,7 +81,9 @@ class ZipSeriesRequestFixedSize extends HttpRequestWithOutput<Set<String>> {
 
     @Override
     protected void cleanup() {
-        temporaryZipFile.delete();
+        if (temporaryZipFile != null) {
+            temporaryZipFile.delete();
+        }
     }
 
     private void writeChunk(long zipFileSize, float zipRatio, FileInputStream fis, OutputStream os) throws IOException {
